user view and controller for add, list of users, details of user

+++ VolunteeringSpringBoot/src/main/java/cs544/mum/edu/repository/UserRepository.java
@@ -1,7 +1,10 @@
 package cs544.mum.edu.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
+
 
 import cs544.mum.edu.domain.User;
 
@@ -10,6 +13,6 @@ import cs544.mum.edu.domain.User;
  *
  */
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-
+public interface UserRepository extends JpaRepository<User, Integer> {
+	
 }
